main file
Getting straight into it. This is all you need to begin your tic tac toe command. This excludes requirements and dependencies, join discord.py (https://dicord.gg/dpy) for help with that. 

This is a basic 2 player tic tac toe command that uses buttons and embeds. This is also a tree (slash) command, I might do a prefix version of this if I'm feeling giddy 👀.